new file runtime-hooks runtime-utils.spec
