+++ ui/src/app/components/apps-panel/expanded-app.tsx
@@ -26,10 +26,10 @@ import { RuntimeStatusIndicator } from 'app/components/runtime-status-indicator'
 import colors from 'app/styles/colors';
 import { reactStyles } from 'app/utils';
 import { setSidebarActiveIconStore } from 'app/utils/navigation';
+import { useRuntimeStatus } from 'app/utils/runtime-hooks';
 import {
   canDeleteRuntime,
   RuntimeStatusRequest,
-  useRuntimeStatus,
 } from 'app/utils/runtime-utils';
 import { runtimeStore, useStore } from 'app/utils/stores';
 import {
